refactor: move tests into more granular files
